requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
